Добавлено разделение доступа к ресурсам в зависимости от роли.

+++ REST-API/Python/quantisk_api/views.py
@@ -1,29 +1,39 @@
-from flask import request, abort
+from flask import request, abort, g
 from flask_restful import Resource
-from .repositories import person_repo, wordpair_repo, site_repo, rank_repo
-from .auth import auth
+from .repositories import person_repo, wordpair_repo, site_repo, rank_repo, user_repo
+from .repositories import NonUniqueError
+from .auth import requires_auth, requires_roles, unauthorized, ADMIN, USER
 import arrow
 from arrow.parser import ParserError
 
 
 class SingleResource(Resource):
     repo = None
-    method_decorators = [auth.login_required]
+    method_decorators = [requires_auth]
 
+    @requires_roles(USER, ADMIN)
     def get(self, id):
         item = self.repo.get_by_id(id)
         if item is None:
             abort(404)
         return item._asdict()
 
+    @requires_roles(ADMIN)
     def put(self, id):
+        item = self.repo.get_by_id(id)
+        if hasattr(item, 'user_id') and not g.user.id == item.user_id:
+            return unauthorized('You can only edit your own stuff')
         body = request.get_json()
         item = self.repo.set(id, **body)
         if item is None:
             abort(404)
         return item._asdict()
 
+    @requires_roles(ADMIN)
     def delete(self, id):
+        item = self.repo.get_by_id(id)
+        if hasattr(item, 'user_id') and not g.user.id == item.user_id:
+            return unauthorized('You can only edit your own stuff')
         if self.repo.delete(id) is None:
             abort(404)
         return None, 204
@@ -31,36 +41,73 @@ class SingleResource(Resource):
 
 class ListResource(Resource):
     repo = None
-    method_decorators = [auth.login_required]
+    method_decorators = [requires_auth]
 
+    @requires_roles(USER, ADMIN)
     def get(self):
         return [i._asdict() for i in self.repo.get_all()]
 
+    @requires_roles(ADMIN)
     def post(self):
         body = request.get_json()
         try:
             item = self.repo.add(**body)
         except TypeError as e:
             abort(400, e)
+        except NonUniqueError as e:
+            abort(409, e)
         else:
             return item._asdict()
+
+class UserResource(SingleResource):
+    repo = user_repo
+
+
+class UserListResource(ListResource):
+    repo = user_repo
 
 
 class PersonResource(SingleResource):
     repo = person_repo
 
-
 class PersonListResource(ListResource):
     repo = person_repo
 
+    @requires_roles(ADMIN)
+    def post(self):
+        body = request.get_json()
+        user_id = g.user.id
+        try:
+            item = self.repo.add(user_id=user_id, **body)
+        except TypeError as e:
+            abort(400, e)
+        except NonUniqueError as e:
+            abort(409, e)
+        else:
+            return item._asdict()
 
 class WordPairResource(SingleResource):
     repo = wordpair_repo
 
 
 class WordPairListResource(ListResource):
+    #Fixme
     repo = wordpair_repo
 
+    @requires_roles(ADMIN)
+    def post(self):
+        body = request.get_json()
+        owner_id = person_repo.get_owner_id(body.person_id)
+        if g.user.id == owner_id:
+            try:
+                item = self.repo.add(**body)
+            except TypeError as e:
+                abort(400, e)
+            except NonUniqueError as e:
+                abort(409, e)
+            else:
+                return item._asdict()
+        return unauthorized('You can only edit your own stuff')
 
 class SiteResource(SingleResource):
     repo = site_repo
@@ -69,11 +116,25 @@ class SiteResource(SingleResource):
 class SiteListResource(ListResource):
     repo = site_repo
 
+    @requires_roles(ADMIN)
+    def post(self):
+        body = request.get_json()
+        user_id = g.user.id
+        try:
+            item = self.repo.add(user_id=user_id, **body)
+        except TypeError as e:
+            abort(400, e)
+        except NonUniqueError as e:
+            abort(409, e)
+        else:
+            return item._asdict()
+
 
 class WordPairsForPersonResource(Resource):
     # Fixme
-    method_decorators = [auth.login_required]
+    method_decorators = [requires_auth]
 
+    @requires_roles(USER, ADMIN)
     def get(self, person_id):
         wordpairs = wordpair_repo.get_by_person_id(person_id)
         if wordpairs is None:
@@ -83,8 +144,9 @@ class WordPairsForPersonResource(Resource):
 
 class TotalRankResource(Resource):
     # Fixme
-    method_decorators = [auth.login_required]
+    method_decorators = [requires_auth]
 
+    @requires_roles(USER, ADMIN)
     def get(self, site_id):
         ranks = rank_repo.get_total(site_id)
         if ranks is None:
@@ -94,8 +156,9 @@ class TotalRankResource(Resource):
 
 class DailyRankResource(Resource):
     # Fixme
-    method_decorators = [auth.login_required]
+    method_decorators = [requires_auth]
 
+    @requires_roles(USER, ADMIN)
     def get(self):
         args = request.args
         person_id = int(args['person_id'])
@@ -107,5 +170,3 @@ class DailyRankResource(Resource):
             return [rank._asdict() for rank in ranks]
         except ParserError:
             abort(400)
-
-
